fix: find fastqs <sample/input>_S[0-9]{1,2}_R{1,2}

+++ tinyMapper.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 
-VERSION=0.11.2
+VERSION=0.11.3
 
 INVOC=$(printf %q "$BASH_SOURCE")$((($#)) && printf ' %q' "$@")
 HASH=`LC_CTYPE=C tr -dc 'A-Z0-9' < /dev/urandom | head -c 6`
@@ -473,6 +473,7 @@ fi
 #       "${SAMPLE}_nxq_R1.fq.gz"
 #       "${SAMPLE}.end1.fq.gz"
 #       "${SAMPLE}.end1.gz"
+#       "${SAMPLE}_S[0-9]{1,2}_R1*.gz"
 if test ! -f "${SAMPLE_R1}" || test ! -f "${SAMPLE_R2}" ; then
     SAMPLE_R1="${SAMPLE}_R1.fastq.gz"
     SAMPLE_R2="${SAMPLE}_R2.fastq.gz"
@@ -498,11 +499,18 @@ if test ! -f "${SAMPLE_R1}" || test ! -f "${SAMPLE_R2}" ; then
                     fn_warning "Sample files found here: ${SAMPLE_R1} & ${SAMPLE_R2}" 2>&1 | tee -a "${LOGFILE}" 
                     fn_warning "Renaming '\${SAMPLE_R1}' & '\${SAMPLE_R2}' variables" 2>&1 | tee -a "${LOGFILE}" 
                 else
-                    fn_error "Sample files not found. Check sample directory: ${SAMPLE_DIR}/." 2>&1 | tee -a "${LOGFILE}"
-                    fn_error "Files *must* be named as follows: ${SAMPLE_BASE}_R1.fq.gz & ${SAMPLE_BASE}_R2.fq.gz" 2>&1 | tee -a "${LOGFILE}"
-                    fn_error "Aborting now." 2>&1 | tee -a "${LOGFILE}"
-                    rm --force "${LOGFILE}"
-                    exit 1
+                    SAMPLE_R1=`find $SAMPLE_DIR -name $SAMPLE_BASE* | grep "_S[0-9]\{1,2\}_R1"`
+                    SAMPLE_R2=`find $SAMPLE_DIR -name $SAMPLE_BASE* | grep "_S[0-9]\{1,2\}_R2"`
+                    if [[ -f "${SAMPLE_R1}" && -f "${SAMPLE_R2}" && "${SAMPLE_R1}" != *[[:space:]]* && "${SAMPLE_R2}" != *[[:space:]]* ]]
+                        fn_warning "Sample files found here: ${SAMPLE_R1} & ${SAMPLE_R2}" 2>&1 | tee -a "${LOGFILE}" 
+                        fn_warning "Renaming '\${SAMPLE_R1}' & '\${SAMPLE_R2}' variables" 2>&1 | tee -a "${LOGFILE}" 
+                    else 
+                        fn_error "Sample files not found. Check sample directory: ${SAMPLE_DIR}/." 2>&1 | tee -a "${LOGFILE}"
+                        fn_error "Files *must* be named as follows: ${SAMPLE_BASE}_R1.fq.gz & ${SAMPLE_BASE}_R2.fq.gz" 2>&1 | tee -a "${LOGFILE}"
+                        fn_error "Aborting now." 2>&1 | tee -a "${LOGFILE}"
+                        rm --force "${LOGFILE}"
+                        exit 1
+                    fi
                 fi
             fi
         fi
@@ -539,37 +547,45 @@ fi
 #       "${INPUT}_R1.fastq.gz"
 #       "${INPUT}_nxq_R1.fq.gz"
 #       "${INPUT}.end1.gz"
+#       "${INPUT}_S[0-9]{1,2}_R1*.gz"
 if test "${DO_INPUT}" == 0 ; then
     if test ! -f "${INPUT_R1}" || test ! -f "${INPUT_R2}" ; then
         INPUT_R1="${INPUT}_R1.fastq.gz"
         INPUT_R2="${INPUT}_R2.fastq.gz"
         if test -f "${INPUT_R1}" && test -f "${INPUT_R2}" ; then
-            fn_warning "Sample files found here: ${INPUT_R1} & ${INPUT_R2}" 2>&1 | tee -a "${LOGFILE}" 
+            fn_warning "Input files found here: ${INPUT_R1} & ${INPUT_R2}" 2>&1 | tee -a "${LOGFILE}" 
             fn_warning "Renaming '\${INPUT_R1}' & '\${INPUT_R2}' variables" 2>&1 | tee -a "${LOGFILE}" 
         else
             INPUT_R1="${INPUT}_nxq_R1.fq.gz"
             INPUT_R2="${INPUT}_nxq_R2.fq.gz"
             if test -f "${INPUT_R1}" && test -f "${INPUT_R2}" ; then
-                fn_warning "Sample files found here: ${INPUT_R1} & ${INPUT_R2}" 2>&1 | tee -a "${LOGFILE}" 
+                fn_warning "Input files found here: ${INPUT_R1} & ${INPUT_R2}" 2>&1 | tee -a "${LOGFILE}" 
                 fn_warning "Renaming '\${INPUT_R1}' & '\${INPUT_R2}' variables" 2>&1 | tee -a "${LOGFILE}" 
             else
                 INPUT_R1="${INPUT}.end1.fq.gz"
                 INPUT_R2="${INPUT}.end2.fq.gz"
                 if test -f "${INPUT_R1}" && test -f "${INPUT_R2}" ; then
-                    fn_warning "Sample files found here: ${INPUT_R1} & ${INPUT_R2}" 2>&1 | tee -a "${LOGFILE}" 
+                    fn_warning "Input files found here: ${INPUT_R1} & ${INPUT_R2}" 2>&1 | tee -a "${LOGFILE}" 
                     fn_warning "Renaming '\${INPUT_R1}' & '\${INPUT_R2}' variables" 2>&1 | tee -a "${LOGFILE}" 
                 else
                     INPUT_R1="${INPUT}.end1.gz"
                     INPUT_R2="${INPUT}.end2.gz"
                     if test -f "${INPUT_R1}" && test -f "${INPUT_R2}" ; then
-                        fn_warning "Sample files found here: ${INPUT_R1} & ${INPUT_R2}" 2>&1 | tee -a "${LOGFILE}" 
+                        fn_warning "Input files found here: ${INPUT_R1} & ${INPUT_R2}" 2>&1 | tee -a "${LOGFILE}" 
                         fn_warning "Renaming '\${INPUT_R1}' & '\${INPUT_R2}' variables" 2>&1 | tee -a "${LOGFILE}" 
                     else
-                        fn_error "Input files are missing. Check input directory: ${INPUT_DIR}/." 2>&1 | tee -a "${LOGFILE}"
-                        fn_error "Files *must* be named as follows: ${INPUT_BASE}_R1.fq.gz & ${INPUT_BASE}_R2.fq.gz" 2>&1 | tee -a "${LOGFILE}"
-                        fn_error "Aborting now." 2>&1 | tee -a "${LOGFILE}"
-                        rm --force "${LOGFILE}"
-                        exit 1
+                        INPUT_R1=`find $INPUT_DIR -name $INPUT_BASE* | grep "_S[0-9]\{1,2\}_R1"`
+                        INPUT_R2=`find $INPUT_DIR -name $INPUT_BASE* | grep "_S[0-9]\{1,2\}_R2"`
+                        if [[ -f "${INPUT_R1}" && -f "${INPUT_R2}" && "${INPUT_R1}" != *[[:space:]]* && "${INPUT_R2}" != *[[:space:]]* ]]
+                            fn_warning "Input files found here: ${INPUT_R1} & ${INPUT_R2}" 2>&1 | tee -a "${LOGFILE}" 
+                            fn_warning "Renaming '\${INPUT_R1}' & '\${INPUT_R2}' variables" 2>&1 | tee -a "${LOGFILE}" 
+                        else 
+                            fn_error "Input files are missing. Check input directory: ${INPUT_DIR}/." 2>&1 | tee -a "${LOGFILE}"
+                            fn_error "Files *must* be named as follows: ${INPUT_BASE}_R1.fq.gz & ${INPUT_BASE}_R2.fq.gz" 2>&1 | tee -a "${LOGFILE}"
+                            fn_error "Aborting now." 2>&1 | tee -a "${LOGFILE}"
+                            rm --force "${LOGFILE}"
+                            exit 1
+                        fi
                     fi
                 fi
             fi
